+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Untitled Config Data Builder
 
+## [1.0.7] - 2022-03-01
+
+- Add support for trailing separators.
+
+- Add support for Vector2Int (int2) and Vector3Int (int3).
+
 ## [1.0.6] - 2022-02-28
 
-Add support for `IMultiSegConfigValueConverter` and multiple separators.
+- Add support for `IMultiSegConfigValueConverter` and multiple separators.
 
 ## [1.0.5] - 2022-02-10
 
-Init.
+- Init.